fix(clientes): correct getClientsCount endpoint in src/controllers/clientesController.js  to query COUNT and return proper JSON/404 handling
