+++ src/controllers/userController.js
@@ -36,36 +36,6 @@ const getUsersInfo = async (req, res) => {
     }
 };
 
-/**
- * Obtiene el número total de usuarios con el rol de 'cliente'.
- */
-const getClientsCount = async (req, res) => {
-    try {
-      // Esta consulta une las tablas usuarios, usuarios_roles y roles
-      // para contar solo aquellos usuarios que tengan el rol de 'cliente'.
-      // Asegúrate de que 'cliente' sea el nombre exacto del rol en tu tabla 'roles'.
-      const query = `
-        SELECT COUNT(u.id) 
-        FROM usuarios u
-        JOIN usuarios_roles ur ON u.id = ur.usuario_id
-        JOIN roles r ON ur.rol_id = r.id
-        WHERE r.rol = 'cliente';
-      `;
-      
-      const result = await pool.query(query);
-  
-      const count = parseInt(result.rows[0].count, 10);
-  
-      res.json({ total_clientes: count });
-  
-    } catch (error) {
-      console.error('Error en getClientsCount:', error);
-      res.status(500).json({ message: 'Error del servidor' });
-    }
-  };
-  
-
 module.exports = {
-    getUsersInfo,
-    getClientsCount
+    getUsersInfo
 };
