+++ src/routes/userRoutes.js
@@ -1,9 +1,9 @@
 const express = require('express');
 const router = express.Router();
 
-const { getUsersInfo, getClientsCount } = require('../controllers/userController');
+const { getUsersInfo } = require('../controllers/userController');
 
 router.get('/users_info', getUsersInfo)
-router.get('/clients_count', getClientsCount)
+
 
 module.exports = router;
